Move prepareGoOptions to utils, add goOptions to gas-oracle, export functions for tests
